feat(fs): implement fat16 cluster writing and harden sys_write with elf size optimization
- Hardened sys_write via vmm_is_mapped validation and Kernel Heap bounce buffers.
- Added 'write' command to the userspace shell for file generation and modification.
- Appended -N (--nmagic) to USER_LDFLAGS to merge ELF text/data segments.
- Eliminated 4KB page padding, reducing photon.bin from 149KB to 96KB to fit boot limits.

+++ src/drivers/ata.c
@@ -184,7 +184,7 @@ int ata_write_sectors(uint32_t lba, uint8_t sector_count,
     return ata_wait_ready();
 }
 
-static size_t ata_block_read(vfs_node_t *node, size_t offset, size_t size,
+static int ata_block_read(vfs_node_t *node, uint64_t offset, uint32_t size,
     uint8_t *buffer)
 {
     (void)node;
@@ -197,7 +197,7 @@ static size_t ata_block_read(vfs_node_t *node, size_t offset, size_t size,
         return 0;
     }
 
-    return size;
+    return (int)size;
 }
 
 static size_t ata_block_write(vfs_node_t *node, size_t offset, size_t size,
